pyproject.toml
# Copyright (c) 2024 Emasoft (for atlasvibe modifications and derivative work)
# Copyright (c) 2024 Atlasvibe (for the original "Atlasvibe Studio" software)
#
# This software is licensed under the MIT License.
# Refer to the LICENSE file for more details.

[project]
name = "atlasvibe"
version = "0.1.0"
description = "General-purpose visual programming IDE for Python with an Agent node flow approach"
authors = [
    {name = "Emasoft", email = "your-email@example.com"}, # Please update with your actual email
]
readme = "README.md"
license = {text = "MIT"}
requires-python = "==3.12.9"
dependencies = [
    "uvicorn[standard]>=0.31.0,<0.32.0",
    "fastapi>=0.104.1,<0.112.0",
    "pyyaml>=6.0.1,<7.0.0",
    "python-dotenv",
    "requests>=2.31.0,<3.0.0",
    "debugpy>=1.8.0,<2.0.0",
    "watchfiles>=0.21.0,<0.22.0",
    "numpy>=1.23.0,<2.0.0",
    "pandas>=2.2.0,<3.0.0",
    "python-box>=7.0.0,<8.0.0",
    "plotly>=5.0.0,<6.0.0",
    "prefect>=3.1.4",
    "chardet>=3.0.0,<6.0.0",
    "griffe",
    "pydantic>=2.11.4",
    "striprtf>=0.0.29",
    "isbinary>=1.0.1",
    "python-multipart>=0.0.9",
    "pathspec>=0.12.1",
    "pytest>=7.4.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2,<8.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "pytest-mock>=3.11.0,<4.0.0",
    "ruff>=0.4.4,<0.5.0",
    "typer[all]",
    "python-frontmatter>=1.0.1,<2.0.0",
    "httpx>=0.26.0,<0.27.0",
    "Pillow>=9.0.0,<11.0.0",      
    "scikit-image>=0.19.0,<0.24.0", 
    "boto3>=1.20.0,<2.0.0",      
    "scipy>=1.7.0,<2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "atlasvibe_engine",
  "pkgs/atlasvibe_sdk",
  "pkgs/atlasvibe",
  "cli",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-rsxX -l --tb=short --strict-markers"
testpaths = [
    "tests",
    "PYTHON/tests",
    "blocks",
    "atlasvibe_engine/tests",
    "cli",
    "playwright-test/fixtures/custom-sequences",
]
python_files = "test_*.py *_test_.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
pythonpath = ["."]
